requirements.txt
fastapi==0.135.3
alembic==1.18.4
pydantic==2.12.5
uvicorn==0.41.0
pydantic-extra-types==2.11.0
pydantic_settings==2.13.1
asyncpg==0.31.0
email-validator==2.3.0
PyJWT==2.12.1
python-jose==3.5.0
bcrypt==5.0.0